+++ Sources/StreamChat/Controllers/ChannelListController/ChannelListController_Tests.swift
@@ -22,8 +22,12 @@ class ChannelListController_Tests: StressTestCase {
     override func setUp() {
         super.setUp()
         
+        setUp(isLocalStorageEnabled: true)
+    }
+    
+    private func setUp(isLocalStorageEnabled: Bool) {
         env = TestEnvironment()
-        client = ChatClient.mock(isLocalStorageEnabled: true)
+        client = ChatClient.mock(isLocalStorageEnabled: isLocalStorageEnabled)
         query = .init(filter: .in(.members, values: [.unique]))
         controller = ChatChannelListController(query: query, client: client, environment: env.environment)
         controllerCallbackQueueID = UUID()
@@ -146,6 +150,43 @@ class ChannelListController_Tests: StressTestCase {
         
         // Assert the updater is called with the query
         XCTAssertEqual(env.channelListUpdater!.update_queries.first?.filter.filterHash, query.filter.filterHash)
+        // Completion shouldn't be called yet
+        XCTAssertFalse(completionCalled)
+        
+        // Simulate successful update
+        env.channelListUpdater!.update_completion?(nil)
+        // Release reference of completion so we can deallocate stuff
+        env.channelListUpdater!.update_completion = nil
+        
+        // Completion should be called
+        AssertAsync.willBeTrue(completionCalled)
+        // `weakController` should be deallocated too
+        AssertAsync.canBeReleased(&weakController)
+    }
+    
+    func test_synchronize_callsChannelQueryUpdater_inOfflineMode() {
+        setUp(isLocalStorageEnabled: false)
+        
+        let queueId = UUID()
+        controller.callbackQueue = .testQueue(withId: queueId)
+        
+        // Simulate `synchronize` calls and catch the completion
+        var completionCalled = false
+        controller.synchronize { error in
+            XCTAssertNil(error)
+            AssertTestQueue(withId: queueId)
+            completionCalled = true
+        }
+        
+        // Keep a weak ref so we can check if it's actually deallocated
+        weak var weakController = controller
+        
+        // (Try to) deallocate the controller
+        // by not keeping any references to it
+        controller = nil
+        
+        // Assert the updater is called with the query
+        XCTAssertEqual(env.channelListUpdater?.update_queries.first?.filter.filterHash, query.filter.filterHash)
         // Completion shouldn't be called yet
         XCTAssertFalse(completionCalled)